requirements.txt
fastapi>=0.104.0
uvicorn>=0.23.2
pydantic>=2.4.2
pyyaml>=6.0.1
langchain>=0.0.335
langchain-openai>=0.0.2
langgraph>=0.0.20
pymongo>=4.5.0
sqlalchemy>=2.0.22
python-dotenv>=1.0.0
tavily-python>=0.2.2
langchain-community>=0.0.10
gradio>=5.0.0
requests>=2.31.0
langchain-chroma>=0.1.0
pypdf>=3.17.0
docx2txt>=0.8
unstructured>=0.10.0
python-magic>=0.4.27
